+++ builtin/logical/transit/backend.go
@@ -6,7 +6,7 @@ import (
 )
 
 func Factory(conf *logical.BackendConfig) (logical.Backend, error) {
-	b := Backend()
+	b := Backend(conf)
 	be, err := b.Backend.Setup(conf)
 	if err != nil {
 		return nil, err
@@ -15,7 +15,7 @@ func Factory(conf *logical.BackendConfig) (logical.Backend, error) {
 	return be, nil
 }
 
-func Backend() *backend {
+func Backend(conf *logical.BackendConfig) *backend {
 	var b backend
 	b.Backend = &framework.Backend{
 		Paths: []*framework.Path{
@@ -33,14 +33,12 @@ func Backend() *backend {
 		Secrets: []*framework.Secret{},
 	}
 
-	b.policies = policyCache{
-		cache: map[string]*lockingPolicy{},
-	}
+	b.lm = newLockManager(conf.System.CachingDisabled())
 
 	return &b
 }
 
 type backend struct {
 	*framework.Backend
-	policies policyCache
+	lm *lockManager
 }