+++ vault/core.go
@@ -215,6 +215,9 @@ type Core struct {
 	maxLeaseTTL     time.Duration
 
 	logger *log.Logger
+
+	// cachingDisabled indicates whether caches are disabled
+	cachingDisabled bool
 }
 
 // CoreConfig is used to parameterize a core
@@ -312,6 +315,7 @@ func NewCore(conf *CoreConfig) (*Core, error) {
 		logger:          conf.Logger,
 		defaultLeaseTTL: conf.DefaultLeaseTTL,
 		maxLeaseTTL:     conf.MaxLeaseTTL,
+		cachingDisabled: conf.DisableCache,
 	}
 
 	// Setup the backends